+++ dev-resources/com/walmartlabs/lacinia/test_utils.clj
@@ -47,9 +47,9 @@
   ;; Return a cleanup fn:
   #(swap! *ping-cleanups inc))
 
-(defn ^:private make-service
+(defn make-service
   "The special option :indirect-schema wraps the schema in a function; this exercises some
-  logic that allows a compiled schema to actually be a  function that returns the compiled schema."
+  logic that allows a compiled schema to actually be a function that returns the compiled schema."
   [options options-builder]
   (let [schema (-> (io/resource "sample-schema.edn")
                    slurp
@@ -100,6 +100,12 @@
          (= method :post)
          (assoc-in [:headers "Content-Type"] "application/graphql")
 
+         (= method :post-json)
+         (->
+           (assoc-in [:headers "Content-Type"] "application/json")
+           (assoc :method :post
+                  :body query))
+
          ;; :post-bad is like :post, but without setting the content type
          (#{:post :post-bad} method)
          (assoc :body query
@@ -118,12 +124,29 @@
   ([method json]
    (send-json-request method json "application/json; charset=utf-8"))
   ([method json content-type]
+   (send-json-request method "/graphql" json content-type))
+  ([method path json content-type]
    (-> {:method method
-        :url "http://localhost:8888/graphql"
+        :url (str "http://localhost:8888" path)
         :throw-exceptions false
         :headers {"Content-Type" content-type}}
        (cond->
          json (assoc :body (cheshire/generate-string json)))
+       client/request
+       (update :body
+               #(try
+                  (cheshire/parse-string % true)
+                  (catch Exception t %))))))
+
+(defn send-json-string-request
+  ([method json]
+   (send-json-string-request method json "application/json; charset=utf-8"))
+  ([method json content-type]
+   (-> {:method method
+        :url "http://localhost:8888/graphql"
+        :throw-exceptions false
+        :headers {"Content-Type" content-type}
+        :body json}
        client/request
        (update :body
                #(try